sgalasso [ia/bots] : player facing

+++ includes/wolf_struct.h
@@ -282,7 +282,7 @@ struct					s_bot // a fusionner avec s_character/ennemies
 	int         init_dir;
 	t_point     position;
 	int         direction;
-	int			player_dist;
+	double		player_dist;
 	int			player_angl;
 	int         detected;	// if it saw player
 	int			alerted;	// if it heard player
